requirements.txt
discord
discord.py
python-dotenv
psycopg2-binary
asyncpg